Create ItemsContainer component and update styling
