+++ src/components/AdminPanel/Skills/UpdateSkill.tsx
@@ -72,7 +72,7 @@ export const UpdateSkill = ({
   ];
 
   return (
-    <Card className="relative w-1/3 h-auto overflow-auto bg-black mt-10 mb-10 border border-white/20">
+    <Card className="absolute w-1/3 h-auto overflow-auto bg-black mt-10 mb-10 border border-white/20">
       <CardHeader>
         <h2 className="text-2xl font-bold text-white">
           {defaultData ? "Update skill" : "Add new skill"}
